+++ src/Client.php
@@ -9,6 +9,7 @@ use Hitmeister\Component\Api\Namespaces\ClaimsNamespace;
 use Hitmeister\Component\Api\Namespaces\ImportFilesNamespace;
 use Hitmeister\Component\Api\Namespaces\ItemsNamespace;
 use Hitmeister\Component\Api\Namespaces\OrdersNamespace;
+use Hitmeister\Component\Api\Namespaces\OrderInvoicesNamespace;
 use Hitmeister\Component\Api\Namespaces\OrderUnitsNamespace;
 use Hitmeister\Component\Api\Namespaces\ProductDataNamespace;
 use Hitmeister\Component\Api\Namespaces\ProductDataStatusNamespace;
@@ -36,7 +37,7 @@ use Hitmeister\Component\Api\Transport\Transport;
  */
 class Client
 {
-	const VERSION = '1.31.1';
+	const VERSION = '1.32.0';
 
 	/** @var Transport */
 	private $transport;
@@ -61,6 +62,9 @@ class Client
 
 	/** @var OrdersNamespace */
 	private $ordersNs;
+
+	/** @var OrderInvoicesNamespace */
+	private $orderInvoicesNs;
 
 	/** @var OrderUnitsNamespace */
 	private $orderUnitsNs;
@@ -119,6 +123,7 @@ class Client
 		$this->importFilesNs = new ImportFilesNamespace($this->transport);
 		$this->itemsNs = new ItemsNamespace($this->transport);
 		$this->ordersNs = new OrdersNamespace($this->transport);
+		$this->orderInvoicesNs = new OrderInvoicesNamespace($this->transport);
 		$this->orderUnitsNs = new OrderUnitsNamespace($this->transport);
 		$this->productDataNs = new ProductDataNamespace($this->transport);
 		$this->productDataStatusNs = new ProductDataStatusNamespace($this->transport);
@@ -197,6 +202,14 @@ class Client
 	public function orders()
 	{
 		return $this->ordersNs;
+	}
+
+	/**
+	 * @return OrderInvoicesNamespace
+	 */
+	public function orderInvoices()
+	{
+		return $this->orderInvoicesNs;
 	}
 
 	/**