+++ CHANGELOG.md
@@ -2,12 +2,23 @@
 
 All Notable changes to `Hitmeister - API SDK for PHP` will be documented in this file.
 
+### 1.32.0 - 2019-08-19
+
+### Added
+
+- Added the following endpoints:
+    - GET `/order-invoices/seller`
+    - GET `/order-invoices/{id_order_invoice}/`
+    - POST `/order-invoices/`
+    - DELETE `/order-invoices/{id_order_invoice}/`
+- Added embedded Field `order_invoices` to `/orders/{id_order}` endpoint
+
+    
 ## 1.31.1 - 2019-08-15
 
 ### Changed
 
 - Allow adding shipment information (tracking numbers) for already marked as send order units via `POST /shipments/`.
-
 
 ## 1.30.1 - 2019-07-19